Add functions to guess version for hotfix/release
Add functions to guess new version number for hotfix and release
according to the Semantic Versioning (http://semver.org/).

This is done by getting the latest git tag and incrementing the MINOR
number and setting to zero the PATCH number for releases.
For hotfixes simply the PATCH number is increased.