Change fix to be closer to the original code
Change the fix to actually use tag_prefix in git tags to be closer to
the original code by just adding an empty string if tagPrefix() does not
return a string (i.e. there is no tag prefix). This way release_name can
be a const again as it gets concated to tag_prefix.